fix: darwin and windows inconsistensicees
